add and view - works!
