+++ frontend/apps/ui/src/features/roles/components/NewRoleModal.tsx
@@ -4,6 +4,7 @@ import {useAddNewRoleMutation} from "@/features/roles/apiSlice"
 import {useTranslation} from "react-i18next"
 import {RoleFormModal} from "kommon"
 import {CheckedNodeStatus} from "@mantine/core"
+import {client2serverPerms} from "@/features/roles/utils"
 
 interface Args {
   opened: boolean
@@ -77,59 +78,4 @@ export default function NewRoleModalContaner({
       onNameChange={onNameChange}
     />
   )
-}
-
-const SCOPE_TO_SKIP = [
-  "folder",
-  "document",
-  "tag",
-  "custom_field",
-  "document_type",
-  "shared_node",
-  "user",
-  "role",
-  "group",
-  "document.page"
-]
-
-const FOLDER_DOCUMENT_NODE_MAP: Record<string, string> = {
-  "folder.view": "node.view",
-  "folder.create": "node.create",
-  "folder.update": "node.update",
-  "folder.move": "node.move",
-  "folder.delete": "node.delete",
-  "document.view": "node.view",
-  "document.update": "node.update",
-  "document.move": "node.move",
-  "document.delete": "node.delete"
-}
-
-const PAGE_MANAGEMENT_MAP: Record<string, string> = {
-  "document.page.extract": "page.extract",
-  "document.page.move": "page.move",
-  "document.page.rotate": "page.rotate",
-  "document.page.reorder": "page.reorder",
-  "document.page.delete": "page.delete"
-}
-
-function client2serverPerms(scopes: string[]): string[] {
-  let result: string[] = []
-
-  scopes.forEach(scope => {
-    if (SCOPE_TO_SKIP.includes(scope)) {
-      //
-    } else {
-      if (Object.keys(FOLDER_DOCUMENT_NODE_MAP).includes(scope)) {
-        result.push(FOLDER_DOCUMENT_NODE_MAP[scope])
-      } else if (Object.keys(PAGE_MANAGEMENT_MAP).includes(scope)) {
-        result.push(PAGE_MANAGEMENT_MAP[scope])
-      } else {
-        result.push(scope)
-      }
-    }
-  })
-
-  result.push("user.me")
-
-  return result
 }
